refactor(http-protocol): restructure announce responses into layered modules
Replaced flat `announce.rs` with `announce/{data,encoding}.rs` directory module.

- `data.rs` — DTO layer: `AnnounceData`, `AnnouncePolicy`, `SwarmMetadata`, `Peer`
- `encoding.rs` — Encoding layer: `Announce<E>`, `Normal`, `Compact`, peers, tests
- `mod.rs` — Re-exports for backward compatibility (no call site changes needed)

All linters and tests pass. Progress: T10 done, spec updated.
