+++ packages/http-protocol/src/v1/responses/announce/data.rs
@@ -1,0 +1,57 @@
+//! DTO (Data Transfer Object) types for the HTTP tracker announce response.
+//!
+//! These are transport-agnostic types describing *what* data goes in the response,
+//! without any encoding logic. They use domain-friendly types (`PeerId`, `SocketAddr`).
+use std::net::SocketAddr;
+
+use derive_more::Constructor;
+use torrust_peer_id::PeerId;
+
+// Protocol-local announce response DTOs intentionally duplicate some domain
+// field shapes. This keeps protocol crates decoupled from tracker domain types
+// and centralizes conversions in boundary adapters.
+#[derive(Clone, Debug, PartialEq, Constructor, Default)]
+pub struct AnnounceData {
+    pub peers: Vec<Peer>,
+    pub stats: SwarmMetadata,
+    pub policy: AnnouncePolicy,
+}
+
+#[derive(PartialEq, Eq, Debug, Clone, Copy, Constructor)]
+pub struct AnnouncePolicy {
+    pub interval: u32,
+    pub interval_min: u32,
+}
+
+impl Default for AnnouncePolicy {
+    fn default() -> Self {
+        Self {
+            interval: 120,
+            interval_min: 120,
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
+pub struct SwarmMetadata {
+    pub complete: u32,
+    pub downloaded: u32,
+    pub incomplete: u32,
+}
+
+impl SwarmMetadata {
+    #[must_use]
+    pub const fn new(complete: u32, downloaded: u32, incomplete: u32) -> Self {
+        Self {
+            complete,
+            downloaded,
+            incomplete,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct Peer {
+    pub peer_id: PeerId,
+    pub peer_addr: SocketAddr,
+}
